+++ app/Http/Controllers/Client/ProjectController.php
@@ -2,8 +2,14 @@
 
 namespace App\Http\Controllers\Client;
 
+use App\Models\Category;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
+
+use function Laravel\Prompts\select;
+use function Ramsey\Uuid\v1;
 
 class ProjectController
 {
@@ -12,9 +18,11 @@ class ProjectController
      */
     public function index(Request $request)
     {
-        $projects = $request->user()->projects()->with('reviews')->get();
+        $projects = Project::with('category')
+        ->where('client_id', $request->user()->id)
+        ->get();
 
-        return $projects;
+        return view('client.projects.index')->with('projects', $projects);
     }
 
     /**
@@ -22,7 +30,7 @@ class ProjectController
      */
     public function create()
     {
-        //
+        return view('client.projects.edit')->with('project', new Project())->with('categories', Category::all());
     }
 
     /**
@@ -30,9 +38,13 @@ class ProjectController
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Project::class);
+
         $request->validate(Project::rules($request->user()->id));
 
-        return $request->user()->projects()->create($request->only(Project::getFillable()));
+        $project = $request->user()->projects()->create($request->only($this->getFillable()));
+
+        return redirect(route('projects.show', $project));
     }
 
     /**
@@ -40,7 +52,8 @@ class ProjectController
      */
     public function show(Project $project)
     {
-        return $project->with('reviews');
+        $project = Project::with('category')->where('id', $project->id)->first();
+        return view('client.projects.show')->with('project', $project);
     }
 
     /**
@@ -48,7 +61,7 @@ class ProjectController
      */
     public function edit(Project $project)
     {
-        return $project;
+        return view('client.projects.edit')->with('project', $project)->with('categories', Category::all());
     }
 
     /**
@@ -56,11 +69,13 @@ class ProjectController
      */
     public function update(Request $request, Project $project)
     {
+        Gate::authorize('update', $project);
+
         $request->validate(Project::rules($request->user()->id, $project->id));
 
-        $project->update($request->only(Project::getFillable()));
+        $project->update($request->only($this->getFillable()));
 
-        return $project;
+        return redirect(route('projects.show', $project));
     }
 
     /**
@@ -68,8 +83,20 @@ class ProjectController
      */
     public function destroy(Project $project)
     {
+        Gate::authorize('delete', $project);
+        
         $project->delete();
 
-        return response()->json(['message' => 'Project deleted successfully'], 200);
+        return redirect(route('projects.index'));
+    }
+
+    private function getFillable(){
+        return [
+            'title',
+            'description',
+            'budget',
+            'deadline',
+            'category_id',
+        ];
     }
 }
