+++ app/Http/Controllers/Freelancer/Projects/ReviewController.php
@@ -4,80 +4,36 @@ namespace App\Http\Controllers\Freelancer\Projects;
 
 use App\Models\Review;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ReviewController
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index(Request $request)
-    {
-        $reviews = $request->user()->workreviews()->get();
-
-        return $reviews;
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
+    
     public function store(Request $request)
     {
-        $request->merge(['review_type' => 'freelancer']);
+        Gate::authorize('create', Review::class);
+
+        $request->merge(['review_type' => 'freelancer', 'project_id' => $request->project_id, 'client_id' => $request->client_id]);
 
         $request->validate(Review::rules($request->client_id, $request->user()->id, $request->project_id));
 
-        $review = $request->user()->workreviews()->create($request->only(Review::getFillable()));
+        $review = new Review();
+
+        $review->fill($request->only(Review::getFillableFields()));
 
         $review->project()->associate($request->project_id);
         $review->client()->associate($request->client_id);
-        $review->save();
+        $request->user()->workreviews()->save($review);
 
-        return $review;
+        return redirect(route('freelancer.projects.index'));
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Review $review)
-    {
-        return $review;
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Review $review)
-    {
-        return $review;
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Review $review)
-    {
-        $request->validate(Review::rules($request->client_id, $request->user()->id, $request->project_id, $review->id));
-
-        $review->update($request->only(Review::getFillable()));
-
-        return $review;
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Review $review)
     {
+        Gate::authorize('delete', $review);
+
         $review->delete();
 
-        return response()->json(['message' => 'Review deleted successfully'], 200);
+        return redirect(route('freelancer.projects.index'));
     }
 }
